Merge dev in prod

+++ frontend/src/services/requests/team.service.ts
@@ -1,22 +1,22 @@
 import api from "../api";
 
 export const createTeam = async (teamName: string, members: number[]) => {
-      const response = await api.post(`team/create`, { teamName, members });
+      const response = await api.post(`team/user/create`, { teamName, members });
       return response.data;
 };
 
 export const getAllTeams = async () => {
-  const res = await api.get("/team/teams");
+  const res = await api.get("/team/admin/teams");
   return res.data.data;
 };
 
 export const getTeamUsers = async (teamId: number) => {
-  const res = await api.get(`/team/teamusers`, { params: { teamId } });
+  const res = await api.get(`/team/admin/teamusers`, { params: { teamId } });
   return res.data.data;
 };
 
 export const getTeamFaction = async (teamId: number) => {
-      const res = await api.get(`/team/teamfaction`, { params: { teamId } });
+      const res = await api.get(`/team/admin/teamfaction`, { params: { teamId } });
       return res.data.data;
 };
 
@@ -27,22 +27,22 @@ export const updateTeam = async (payload: {
   teamMembers: number[] | null;
   type: string | null;
 }) => {
-  const res = await api.put("/team/modify", payload);
+  const res = await api.put("/team/admin/modify", payload);
   return res.data;
 };
 
 export const deleteTeam = async (teamID: number) => {
-  const res = await api.delete("/team/delete", { params: { teamID } });
+  const res = await api.delete("/team/admin/delete", { params: { teamID } });
   return res.data;
 };
 
 export const createTeamLight = async (payload: { teamName: string; factionId: number | null }) => {
-  const res = await api.post("/team/createlight", payload);
+  const res = await api.post("/team/admin/createlight", payload);
   return res.data;
 };
 
 export const teamDistribution = async () => {
-      const response = await api.post("team/distributeteam");
+      const response = await api.post("team/admin/distributeteam");
       return response.data
 };
